# tools/sdk-ts/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Notifo API",
    "version": "1.0.0",
    "x-logo": {
      "url": "/logo.svg"
    }
  },
  "servers": [
    {
      "url": "https://localhost:5002"
    }
  ],
  "paths": {
    "/api/me": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get the current user.",
        "operationId": "User_GetUser",
        "responses": {
          "200": {
            "description": "User returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getUser"
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Update the user.",
        "operationId": "User_PostUser",
        "requestBody": {
          "x-name": "request",
          "description": "The upsert request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "Users upserted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "postUser"
      }
    },
    "/api/me/token": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get the token for the internal notifo app.",
        "operationId": "User_GetInternalToken",
        "responses": {
          "200": {
            "description": "User returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getInternalToken"
      }
    },
    "/api/me/topics": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Query the user topics.",
        "operationId": "User_GetTopics",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "description": "The optional language.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "User subscriptions returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserTopicDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getTopics"
      }
    },
    "/api/me/subscriptions": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Query the user subscriptions.",
        "operationId": "User_GetMySubscriptions",
        "parameters": [
          {
            "name": "Topics",
            "in": "query",
            "description": "The topics we are interested in.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "query",
            "in": "query",
            "description": "The optional query to search for items.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "take",
            "in": "query",
            "description": "The number of items to return.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "skip",
            "in": "query",
            "description": "The number of items to skip.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "User subscriptions returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseDtoOfSubscriptionDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getMySubscriptions"
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Upserts or deletes my subscriptions.",
        "description": "User Id and App Id are resolved using the API token.",
        "operationId": "User_PostMySubscriptions",
        "requestBody": {
          "x-name": "request",
          "description": "The subscription settings.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeManyDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "204": {
            "description": "User subscribed."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "postMySubscriptions"
      }
    },
    "/api/me/subscriptions/{topic}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Gets a user subscription.",
        "description": "User Id and App Id are resolved using the API token.",
        "operationId": "User_GetMySubscription",
        "parameters": [
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "description": "The topic path.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDto"
                }
              }
            }
          },
          "404": {
            "description": "Subscription does not exist."
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getMySubscription"
      }
    },
    "/api/me/subscriptions/{prefix}": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Remove my subscription.",
        "description": "User Id and App Id are resolved using the API token.",
        "operationId": "User_DeleteSubscription",
        "parameters": [
          {
            "name": "prefix",
            "in": "path",
            "required": true,
            "description": "The topic prefix.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": "User unsubscribed."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "deleteSubscription"
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProfileDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "fullName",
          "emailAddress",
          "allowedTopics",
          "supportedLanguages",
          "supportedTimezones",
          "settings"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "description": "The full name of the user."
          },
          "emailAddress": {
            "type": "string",
            "description": "The email of the user."
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number.",
            "nullable": true
          },
          "allowedTopics": {
            "type": "array",
            "description": "The allowed Topics.",
            "items": {
              "type": "string"
            }
          },
          "preferredLanguage": {
            "type": "string",
            "description": "The preferred language of the user.",
            "nullable": true
          },
          "preferredTimezone": {
            "type": "string",
            "description": "The timezone of the user.",
            "nullable": true
          },
          "supportedLanguages": {
            "type": "array",
            "description": "The supported languages.",
            "items": {
              "type": "string"
            }
          },
          "supportedTimezones": {
            "type": "array",
            "description": "The supported timezones.",
            "items": {
              "type": "string"
            }
          },
          "settings": {
            "type": "object",
            "description": "Notification settings per channel.",
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelSettingDto"
            }
          }
        }
      },
      "ChannelSettingDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "send",
          "condition",
          "required"
        ],
        "properties": {
          "send": {
            "description": "Defines if to send a notification through this channel.",
            "$ref": "#/components/schemas/ChannelSend"
          },
          "condition": {
            "description": "Defines when to send a notification through this channel.",
            "$ref": "#/components/schemas/ChannelCondition"
          },
          "required": {
            "description": "Defines the requirements for this channel.",
            "$ref": "#/components/schemas/ChannelRequired"
          },
          "delayInSeconds": {
            "type": "integer",
            "description": "The delay in seconds.",
            "format": "int32",
            "nullable": true
          },
          "template": {
            "type": "string",
            "description": "The template if the channel supports it.",
            "nullable": true
          },
          "groupKey": {
            "type": "string",
            "description": "The grouping key to combine notifications per channel.",
            "nullable": true
          },
          "properties": {
            "description": "Additional properties.",
            "nullable": true,
            "$ref": "#/components/schemas/NotificationProperties"
          }
        }
      },
      "ChannelSend": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Inherit",
          "Send",
          "NotSending",
          "NotAllowed"
        ],
        "enum": [
          "Inherit",
          "Send",
          "NotSending",
          "NotAllowed"
        ]
      },
      "ChannelCondition": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Inherit",
          "IfNotSeen",
          "IfNotConfirmed",
          "Always"
        ],
        "enum": [
          "Inherit",
          "IfNotSeen",
          "IfNotConfirmed",
          "Always"
        ]
      },
      "ChannelRequired": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Inherit",
          "NotRequired",
          "Required"
        ],
        "enum": [
          "Inherit",
          "NotRequired",
          "Required"
        ]
      },
      "NotificationProperties": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        }
      },
      "ErrorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "statusCode"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "The error message.",
            "nullable": true
          },
          "errorCode": {
            "type": "string",
            "description": "The error code.",
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "description": "The optional trace id.",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "The error type, usually a link.",
            "nullable": true
          },
          "details": {
            "type": "array",
            "description": "Detailed error messages.",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "statusCode": {
            "type": "integer",
            "description": "Status code of the http response.",
            "format": "int32"
          }
        }
      },
      "AdminProfileDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "description": "The token for the integrated app.",
            "nullable": true
          }
        }
      },
      "UpdateProfileDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "fullName": {
            "type": "string",
            "description": "The full name of the user.",
            "nullable": true
          },
          "emailAddress": {
            "type": "string",
            "description": "The email of the user.",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number.",
            "nullable": true
          },
          "preferredLanguage": {
            "type": "string",
            "description": "The preferred language of the user.",
            "nullable": true
          },
          "preferredTimezone": {
            "type": "string",
            "description": "The timezone of the user.",
            "nullable": true
          },
          "settings": {
            "type": "object",
            "description": "Notification settings per channel.",
            "nullable": true,
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelSettingDto"
            }
          }
        }
      },
      "UserTopicDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "path",
          "name",
          "showAutomatically",
          "channels"
        ],
        "properties": {
          "path": {
            "type": "string",
            "description": "The path."
          },
          "name": {
            "type": "string",
            "description": "The name."
          },
          "description": {
            "type": "string",
            "description": "The optional description.",
            "nullable": true
          },
          "showAutomatically": {
            "type": "boolean",
            "description": "True to show the topic automatically to new users, e.g. when he accepts push notifications."
          },
          "channels": {
            "type": "object",
            "description": "The channel options.",
            "additionalProperties": {
              "$ref": "#/components/schemas/TopicChannel"
            }
          }
        }
      },
      "TopicChannel": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "NotAllowed",
          "Allowed"
        ],
        "enum": [
          "NotAllowed",
          "Allowed"
        ]
      },
      "ListResponseDtoOfSubscriptionDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "description": "The items.",
            "items": {
              "$ref": "#/components/schemas/SubscriptionDto"
            }
          },
          "total": {
            "type": "integer",
            "description": "The total number of items.",
            "format": "int64"
          }
        }
      },
      "SubscriptionDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "topicPrefix",
          "topicSettings"
        ],
        "properties": {
          "topicPrefix": {
            "type": "string",
            "description": "The topic to add."
          },
          "topicSettings": {
            "type": "object",
            "description": "Notification settings per channel.",
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelSettingDto"
            }
          },
          "scheduling": {
            "description": "The scheduling settings.",
            "nullable": true,
            "$ref": "#/components/schemas/SchedulingDto"
          }
        }
      },
      "SchedulingDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "time"
        ],
        "properties": {
          "type": {
            "description": "The scheduling type.",
            "$ref": "#/components/schemas/SchedulingType"
          },
          "nextWeekDay": {
            "description": "To schedule the event at the next day of the week.",
            "nullable": true,
            "$ref": "#/components/schemas/IsoDayOfWeek"
          },
          "date": {
            "type": "string",
            "description": "The scheduling date.",
            "format": "date",
            "nullable": true
          },
          "time": {
            "type": "string",
            "description": "The scheduling time.",
            "format": "time"
          },
          "delayInSeconds": {
            "type": "integer",
            "description": "The delay relative to the server time.",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "SchedulingType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "UTC",
          "UserTime"
        ],
        "enum": [
          "UTC",
          "UserTime"
        ]
      },
      "IsoDayOfWeek": {
        "type": "string",
        "description": "Equates the days of the week with their numerical value according to\nISO-8601. This corresponds with System.DayOfWeek except for Sunday, which\nis 7 in the ISO numbering and 0 in System.DayOfWeek.",
        "x-enumNames": [
          "None",
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Saturday",
          "Sunday"
        ],
        "enum": [
          "None",
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Saturday",
          "Sunday"
        ]
      },
      "SubscribeManyDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "subscribe": {
            "type": "array",
            "description": "A list of topics to create.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/SubscribeDto"
            }
          },
          "unsubscribe": {
            "type": "array",
            "description": "A list of topics to unsubscribe from.",
            "nullable": true,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SubscribeDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "topicPrefix"
        ],
        "properties": {
          "topicPrefix": {
            "type": "string",
            "description": "The topic to add.",
            "minLength": 1
          },
          "topicSettings": {
            "type": "object",
            "description": "Notification settings per channel.",
            "nullable": true,
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelSettingDto"
            }
          },
          "scheduling": {
            "description": "The scheduling settings.",
            "nullable": true,
            "$ref": "#/components/schemas/SchedulingDto"
          },
          "hasScheduling": {
            "type": "boolean",
            "description": "Indicates whether scheduling should be overriden."
          }
        }
      },
...
    }
  }
}